Factor shared test boilerplate into pkcs11_test_util.h
The F-* regression tests each duplicated the module load/unload, the
funcList/dlib globals, the CHECK_RV/CHECK_TRUE macros and the
pass/fail summary. Move them to tests/pkcs11_test_util.h along with a
pkcs11_open_session() helper for the common C_Initialize + slot +
R/W session setup, and use them from the tests.
