+++ tests/create_object_class_test.c
@@ -18,11 +18,8 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
  *
- * Regression test for issues F-5513 and F-5516. CreateObject left objectClass
- * at the (CK_OBJECT_CLASS)-1 sentinel when CKA_CLASS was omitted and then
- * treated the template as a key object, creating an object whose CKA_CLASS is
- * not a valid PKCS#11 class. A complete key template that omits only
- * CKA_CLASS must be rejected with CKR_TEMPLATE_INCOMPLETE.
+ * Regression test for issues F-5513 and F-5516: a C_CreateObject key template
+ * that omits CKA_CLASS must be rejected with CKR_TEMPLATE_INCOMPLETE.
  */
 
 #ifdef HAVE_CONFIG_H
@@ -48,94 +45,25 @@
 #endif
 
 #include "testdata.h"
+#include "pkcs11_test_util.h"
 
 #define TEST_DIR "./store/create_object_class_test"
-
-static int test_passed = 0;
-static int test_failed = 0;
-
-#define CHECK_RV(rv, op, expected) do {                                       \
-    if ((rv) != (expected)) {                                                 \
-        fprintf(stderr, "FAIL: %s: expected 0x%lx, got 0x%lx\n", op,          \
-                (unsigned long)(expected), (unsigned long)(rv));              \
-        test_failed++;                                                        \
-    } else {                                                                  \
-        printf("PASS: %s\n", op);                                             \
-        test_passed++;                                                        \
-    }                                                                         \
-} while (0)
-
-#ifndef HAVE_PKCS11_STATIC
-static void* dlib;
-#endif
-static CK_FUNCTION_LIST* funcList;
-
-static CK_RV pkcs11_load(void)
-{
-    CK_RV ret;
-#ifndef HAVE_PKCS11_STATIC
-    CK_C_GetFunctionList func;
-
-    dlib = dlopen(WOLFPKCS11_DLL_FILENAME, RTLD_NOW | RTLD_LOCAL);
-    if (dlib == NULL) {
-        fprintf(stderr, "dlopen error: %s\n", dlerror());
-        return CKR_GENERAL_ERROR;
-    }
-    func = (CK_C_GetFunctionList)dlsym(dlib, "C_GetFunctionList");
-    if (func == NULL) {
-        fprintf(stderr, "Failed to get function list function\n");
-        dlclose(dlib);
-        return CKR_GENERAL_ERROR;
-    }
-    ret = func(&funcList);
-    if (ret != CKR_OK) {
-        dlclose(dlib);
-        return ret;
-    }
-#else
-    ret = C_GetFunctionList(&funcList);
-    if (ret != CKR_OK)
-        return ret;
-#endif
-    return CKR_OK;
-}
-
-static void pkcs11_unload(void)
-{
-#ifndef HAVE_PKCS11_STATIC
-    if (dlib != NULL) {
-        dlclose(dlib);
-        dlib = NULL;
-    }
-#endif
-    funcList = NULL;
-}
 
 static int run_test(void)
 {
     CK_RV rv;
-    CK_C_INITIALIZE_ARGS args;
-    CK_SLOT_ID slotList[16];
-    CK_ULONG slotCount = sizeof(slotList) / sizeof(slotList[0]);
-    CK_SLOT_ID slot = 0;
     CK_SESSION_HANDLE session = 0;
-    int sessFlags = CKF_SERIAL_SESSION | CKF_RW_SESSION;
     CK_OBJECT_HANDLE obj = CK_INVALID_HANDLE;
     CK_OBJECT_CLASS secretClass = CKO_SECRET_KEY;
     CK_KEY_TYPE genericType = CKK_GENERIC_SECRET;
     CK_BBOOL ckFalse = CK_FALSE;
     byte keyData[16] = { 0 };
-
-    /* Complete key material, but no CKA_CLASS. Pre-fix this created an
-     * object with an invalid (sentinel) class. */
     CK_ATTRIBUTE noClassTmpl[] = {
         { CKA_KEY_TYPE, &genericType, sizeof(genericType) },
         { CKA_VALUE,    keyData,      sizeof(keyData)      },
         { CKA_PRIVATE,  &ckFalse,     sizeof(ckFalse)      },
     };
     CK_ULONG noClassTmplCnt = sizeof(noClassTmpl) / sizeof(*noClassTmpl);
-
-    /* Same template with CKA_CLASS supplied: must succeed. */
     CK_ATTRIBUTE withClassTmpl[] = {
         { CKA_CLASS,    &secretClass, sizeof(secretClass) },
         { CKA_KEY_TYPE, &genericType, sizeof(genericType) },
@@ -149,29 +77,14 @@ static int run_test(void)
     if (rv != CKR_OK)
         return -1;
 
-    XMEMSET(&args, 0, sizeof(args));
-    args.flags = CKF_OS_LOCKING_OK;
-    rv = funcList->C_Initialize(&args);
-    CHECK_RV(rv, "C_Initialize", CKR_OK);
+    rv = pkcs11_open_session(&session);
+    CHECK_RV(rv, "open session", CKR_OK);
     if (rv != CKR_OK)
         goto out;
 
-    rv = funcList->C_GetSlotList(CK_TRUE, slotList, &slotCount);
-    CHECK_RV(rv, "C_GetSlotList", CKR_OK);
-    if (rv != CKR_OK || slotCount == 0)
-        goto out;
-    slot = slotList[0];
-
-    rv = funcList->C_OpenSession(slot, sessFlags, NULL, NULL, &session);
-    CHECK_RV(rv, "C_OpenSession", CKR_OK);
-    if (rv != CKR_OK)
-        goto out;
-
-    /* Key template missing only CKA_CLASS must be rejected. */
     rv = funcList->C_CreateObject(session, noClassTmpl, noClassTmplCnt, &obj);
     CHECK_RV(rv, "C_CreateObject(no CKA_CLASS)", CKR_TEMPLATE_INCOMPLETE);
 
-    /* The same template with CKA_CLASS supplied still succeeds. */
     rv = funcList->C_CreateObject(session, withClassTmpl, withClassTmplCnt,
                                   &obj);
     CHECK_RV(rv, "C_CreateObject(with CKA_CLASS)", CKR_OK);
@@ -195,14 +108,5 @@ int main(int argc, char* argv[])
 
     printf("=== wolfPKCS11 C_CreateObject CKA_CLASS test ===\n");
     run_test();
-
-    printf("\n=== Test Results ===\n");
-    printf("Tests passed: %d\n", test_passed);
-    printf("Tests failed: %d\n", test_failed);
-    if (test_failed == 0)
-        printf("ALL TESTS PASSED!\n");
-    else
-        printf("SOME TESTS FAILED!\n");
-
-    return (test_failed == 0) ? 0 : 1;
+    return pkcs11_test_summary();
 }
