+++ tests/decrypt_final_bufsize_test.c
@@ -18,12 +18,9 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
  *
- * Regression test for issue F-6050. The CKM_AES_CBC_PAD branch of
- * C_DecryptFinal handed WP11_AesCbcPad_DecryptFinal the buffered block length
- * (16) instead of the caller's output-buffer capacity, so the helper's
- * BUFFER_E guard could never fire and an undersized caller buffer was
- * overflowed by up to 15 bytes. C_DecryptFinal must return
- * CKR_BUFFER_TOO_SMALL for a too-small buffer like the GCM/CTS branches.
+ * Regression test for issue F-6050: the CKM_AES_CBC_PAD branch of
+ * C_DecryptFinal must report CKR_BUFFER_TOO_SMALL for an undersized output
+ * buffer instead of overflowing it.
  */
 
 #ifdef HAVE_CONFIG_H
@@ -49,88 +46,14 @@
 #endif
 
 #include "testdata.h"
+#include "pkcs11_test_util.h"
 
 #define TEST_DIR "./store/decrypt_final_bufsize_test"
-
-static int test_passed = 0;
-static int test_failed = 0;
-
-#define CHECK_RV(rv, op, expected) do {                                       \
-    if ((rv) != (expected)) {                                                 \
-        fprintf(stderr, "FAIL: %s: expected 0x%lx, got 0x%lx\n", op,          \
-                (unsigned long)(expected), (unsigned long)(rv));              \
-        test_failed++;                                                        \
-    } else {                                                                  \
-        printf("PASS: %s\n", op);                                             \
-        test_passed++;                                                        \
-    }                                                                         \
-} while (0)
-
-#define CHECK_TRUE(cond, op) do {                                             \
-    if (!(cond)) {                                                            \
-        fprintf(stderr, "FAIL: %s\n", op);                                    \
-        test_failed++;                                                        \
-    } else {                                                                  \
-        printf("PASS: %s\n", op);                                             \
-        test_passed++;                                                        \
-    }                                                                         \
-} while (0)
-
-#ifndef HAVE_PKCS11_STATIC
-static void* dlib;
-#endif
-static CK_FUNCTION_LIST* funcList;
-
-static CK_RV pkcs11_load(void)
-{
-    CK_RV ret;
-#ifndef HAVE_PKCS11_STATIC
-    CK_C_GetFunctionList func;
-
-    dlib = dlopen(WOLFPKCS11_DLL_FILENAME, RTLD_NOW | RTLD_LOCAL);
-    if (dlib == NULL) {
-        fprintf(stderr, "dlopen error: %s\n", dlerror());
-        return CKR_GENERAL_ERROR;
-    }
-    func = (CK_C_GetFunctionList)dlsym(dlib, "C_GetFunctionList");
-    if (func == NULL) {
-        fprintf(stderr, "Failed to get function list function\n");
-        dlclose(dlib);
-        return CKR_GENERAL_ERROR;
-    }
-    ret = func(&funcList);
-    if (ret != CKR_OK) {
-        dlclose(dlib);
-        return ret;
-    }
-#else
-    ret = C_GetFunctionList(&funcList);
-    if (ret != CKR_OK)
-        return ret;
-#endif
-    return CKR_OK;
-}
-
-static void pkcs11_unload(void)
-{
-#ifndef HAVE_PKCS11_STATIC
-    if (dlib != NULL) {
-        dlclose(dlib);
-        dlib = NULL;
-    }
-#endif
-    funcList = NULL;
-}
 
 static int run_test(void)
 {
     CK_RV rv;
-    CK_C_INITIALIZE_ARGS args;
-    CK_SLOT_ID slotList[16];
-    CK_ULONG slotCount = sizeof(slotList) / sizeof(slotList[0]);
-    CK_SLOT_ID slot = 0;
     CK_SESSION_HANDLE session = 0;
-    int sessFlags = CKF_SERIAL_SESSION | CKF_RW_SESSION;
     CK_OBJECT_HANDLE aesKey = CK_INVALID_HANDLE;
     CK_OBJECT_CLASS secretClass = CKO_SECRET_KEY;
     CK_KEY_TYPE aesType = CKK_AES;
@@ -138,8 +61,8 @@ static int run_test(void)
     CK_BBOOL ckFalse = CK_FALSE;
     CK_MECHANISM mech;
     byte iv[16];
-    /* 15-byte plaintext -> one padded block; the held-back final block
-     * unpads to 15 bytes, the maximum, maximising the overflow window. */
+    /* 15-byte plaintext -> one padded block whose final block unpads to 15
+     * bytes, the maximum unpadded size. */
     byte plain[15];
     byte enc[32];
     byte upd[32];
@@ -166,21 +89,8 @@ static int run_test(void)
     if (rv != CKR_OK)
         return -1;
 
-    XMEMSET(&args, 0, sizeof(args));
-    args.flags = CKF_OS_LOCKING_OK;
-    rv = funcList->C_Initialize(&args);
-    CHECK_RV(rv, "C_Initialize", CKR_OK);
-    if (rv != CKR_OK)
-        goto out;
-
-    rv = funcList->C_GetSlotList(CK_TRUE, slotList, &slotCount);
-    CHECK_RV(rv, "C_GetSlotList", CKR_OK);
-    if (rv != CKR_OK || slotCount == 0)
-        goto out;
-    slot = slotList[0];
-
-    rv = funcList->C_OpenSession(slot, sessFlags, NULL, NULL, &session);
-    CHECK_RV(rv, "C_OpenSession", CKR_OK);
+    rv = pkcs11_open_session(&session);
+    CHECK_RV(rv, "open session", CKR_OK);
     if (rv != CKR_OK)
         goto out;
 
@@ -194,7 +104,6 @@ static int run_test(void)
     mech.pParameter = iv;
     mech.ulParameterLen = sizeof(iv);
 
-    /* Encrypt: 15 bytes -> 16 bytes ciphertext. */
     rv = funcList->C_EncryptInit(session, &mech, aesKey);
     CHECK_RV(rv, "C_EncryptInit", CKR_OK);
     if (rv != CKR_OK)
@@ -206,8 +115,8 @@ static int run_test(void)
         goto out;
     CHECK_TRUE(encSz == 16, "ciphertext is one block");
 
-    /* Baseline: a complete multi-part decrypt with an adequate final buffer
-     * still round-trips the plaintext after the fix. */
+    /* Baseline: a full multi-part decrypt with an adequate final buffer
+     * round-trips the plaintext. */
     rv = funcList->C_DecryptInit(session, &mech, aesKey);
     CHECK_RV(rv, "C_DecryptInit (baseline)", CKR_OK);
     if (rv != CKR_OK)
@@ -225,10 +134,9 @@ static int run_test(void)
                XMEMCMP(lastPart, plain, sizeof(plain)) == 0,
                "plaintext round-trips with adequate buffer");
 
-    /* Fresh decrypt that supplies an undersized final buffer: claim only 4
-     * bytes of capacity and guard the rest with a canary. C_DecryptFinal must
-     * return CKR_BUFFER_TOO_SMALL without writing past the 4 bytes. Pre-fix it
-     * wrote up to 15 bytes and returned CKR_OK, overflowing the buffer. */
+    /* Undersized final buffer: claim 4 bytes and guard the rest with a
+     * canary. C_DecryptFinal must report CKR_BUFFER_TOO_SMALL without writing
+     * past the 4 bytes. */
     rv = funcList->C_DecryptInit(session, &mech, aesKey);
     CHECK_RV(rv, "C_DecryptInit (undersized)", CKR_OK);
     if (rv != CKR_OK)
@@ -276,14 +184,5 @@ int main(int argc, char* argv[])
 
     printf("=== wolfPKCS11 C_DecryptFinal CBC-PAD buffer-size test ===\n");
     run_test();
-
-    printf("\n=== Test Results ===\n");
-    printf("Tests passed: %d\n", test_passed);
-    printf("Tests failed: %d\n", test_failed);
-    if (test_failed == 0)
-        printf("ALL TESTS PASSED!\n");
-    else
-        printf("SOME TESTS FAILED!\n");
-
-    return (test_failed == 0) ? 0 : 1;
+    return pkcs11_test_summary();
 }
