+++ tests/include.am
@@ -209,4 +209,5 @@ endif
 
 EXTRA_DIST += tests/unit.h \
               tests/testdata.h \
+              tests/pkcs11_test_util.h \
               tests/README.md
