+++ tests/key_function_not_permitted_test.c
@@ -18,11 +18,9 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
  *
- * Regression test for issue F-6052. CheckOpSupported denied an operation
- * whose CKA_<op> usage attribute was CK_FALSE with CKR_KEY_TYPE_INCONSISTENT.
- * PKCS#11 mandates CKR_KEY_FUNCTION_NOT_PERMITTED for a key whose usage
- * attribute does not allow the operation; CKR_KEY_TYPE_INCONSISTENT is for a
- * key type that does not match the mechanism.
+ * Regression test for issue F-6052: an operation denied by a key's CKA_<op>
+ * usage attribute must report CKR_KEY_FUNCTION_NOT_PERMITTED, distinct from
+ * the key-type mismatch case (CKR_KEY_TYPE_INCONSISTENT).
  */
 
 #ifdef HAVE_CONFIG_H
@@ -48,78 +46,14 @@
 #endif
 
 #include "testdata.h"
+#include "pkcs11_test_util.h"
 
 #define TEST_DIR "./store/key_function_not_permitted_test"
-
-static int test_passed = 0;
-static int test_failed = 0;
-
-#define CHECK_RV(rv, op, expected) do {                                       \
-    if ((rv) != (expected)) {                                                 \
-        fprintf(stderr, "FAIL: %s: expected 0x%lx, got 0x%lx\n", op,          \
-                (unsigned long)(expected), (unsigned long)(rv));              \
-        test_failed++;                                                        \
-    } else {                                                                  \
-        printf("PASS: %s\n", op);                                             \
-        test_passed++;                                                        \
-    }                                                                         \
-} while (0)
-
-#ifndef HAVE_PKCS11_STATIC
-static void* dlib;
-#endif
-static CK_FUNCTION_LIST* funcList;
-
-static CK_RV pkcs11_load(void)
-{
-    CK_RV ret;
-#ifndef HAVE_PKCS11_STATIC
-    CK_C_GetFunctionList func;
-
-    dlib = dlopen(WOLFPKCS11_DLL_FILENAME, RTLD_NOW | RTLD_LOCAL);
-    if (dlib == NULL) {
-        fprintf(stderr, "dlopen error: %s\n", dlerror());
-        return CKR_GENERAL_ERROR;
-    }
-    func = (CK_C_GetFunctionList)dlsym(dlib, "C_GetFunctionList");
-    if (func == NULL) {
-        fprintf(stderr, "Failed to get function list function\n");
-        dlclose(dlib);
-        return CKR_GENERAL_ERROR;
-    }
-    ret = func(&funcList);
-    if (ret != CKR_OK) {
-        dlclose(dlib);
-        return ret;
-    }
-#else
-    ret = C_GetFunctionList(&funcList);
-    if (ret != CKR_OK)
-        return ret;
-#endif
-    return CKR_OK;
-}
-
-static void pkcs11_unload(void)
-{
-#ifndef HAVE_PKCS11_STATIC
-    if (dlib != NULL) {
-        dlclose(dlib);
-        dlib = NULL;
-    }
-#endif
-    funcList = NULL;
-}
 
 static int run_test(void)
 {
     CK_RV rv;
-    CK_C_INITIALIZE_ARGS args;
-    CK_SLOT_ID slotList[16];
-    CK_ULONG slotCount = sizeof(slotList) / sizeof(slotList[0]);
-    CK_SLOT_ID slot = 0;
     CK_SESSION_HANDLE session = 0;
-    int sessFlags = CKF_SERIAL_SESSION | CKF_RW_SESSION;
     CK_OBJECT_HANDLE aesKey = CK_INVALID_HANDLE;
     CK_OBJECT_CLASS secretClass = CKO_SECRET_KEY;
     CK_KEY_TYPE aesType = CKK_AES;
@@ -134,7 +68,6 @@ static int run_test(void)
         /* Correct key type for the mechanism, but usage is denied. */
         { CKA_ENCRYPT,  &ckFalse,     sizeof(ckFalse)     },
         { CKA_DECRYPT,  &ckFalse,     sizeof(ckFalse)     },
-        /* Public object so it is reachable without C_Login. */
         { CKA_PRIVATE,  &ckFalse,     sizeof(ckFalse)     },
     };
     CK_ULONG aesTmplCnt = sizeof(aesTmpl) / sizeof(*aesTmpl);
@@ -144,21 +77,8 @@ static int run_test(void)
     if (rv != CKR_OK)
         return -1;
 
-    XMEMSET(&args, 0, sizeof(args));
-    args.flags = CKF_OS_LOCKING_OK;
-    rv = funcList->C_Initialize(&args);
-    CHECK_RV(rv, "C_Initialize", CKR_OK);
-    if (rv != CKR_OK)
-        goto out;
-
-    rv = funcList->C_GetSlotList(CK_TRUE, slotList, &slotCount);
-    CHECK_RV(rv, "C_GetSlotList", CKR_OK);
-    if (rv != CKR_OK || slotCount == 0)
-        goto out;
-    slot = slotList[0];
-
-    rv = funcList->C_OpenSession(slot, sessFlags, NULL, NULL, &session);
-    CHECK_RV(rv, "C_OpenSession", CKR_OK);
+    rv = pkcs11_open_session(&session);
+    CHECK_RV(rv, "open session", CKR_OK);
     if (rv != CKR_OK)
         goto out;
 
@@ -172,8 +92,6 @@ static int run_test(void)
     mech.pParameter = iv;
     mech.ulParameterLen = sizeof(iv);
 
-    /* Correct key type, but CKA_ENCRYPT is CK_FALSE: the operation is not
-     * permitted, which is distinct from a key-type mismatch. */
     rv = funcList->C_EncryptInit(session, &mech, aesKey);
     CHECK_RV(rv, "C_EncryptInit(CKA_ENCRYPT=FALSE)",
              CKR_KEY_FUNCTION_NOT_PERMITTED);
@@ -201,14 +119,5 @@ int main(int argc, char* argv[])
 
     printf("=== wolfPKCS11 key-function-not-permitted test ===\n");
     run_test();
-
-    printf("\n=== Test Results ===\n");
-    printf("Tests passed: %d\n", test_passed);
-    printf("Tests failed: %d\n", test_failed);
-    if (test_failed == 0)
-        printf("ALL TESTS PASSED!\n");
-    else
-        printf("SOME TESTS FAILED!\n");
-
-    return (test_failed == 0) ? 0 : 1;
+    return pkcs11_test_summary();
 }
