+++ tests/login_another_user_test.c
@@ -18,12 +18,9 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
  *
- * Regression test for issue F-4527. C_Login conflated the two
- * already-logged-in cases: logging in while the *same* user type is logged
- * in must return CKR_USER_ALREADY_LOGGED_IN, while logging in while a
- * *different* user type is logged in must return
- * CKR_USER_ANOTHER_ALREADY_LOGGED_IN. Pre-fix every case returned
- * CKR_USER_ALREADY_LOGGED_IN.
+ * Regression test for issue F-4527: C_Login must distinguish the
+ * same-user-type case (CKR_USER_ALREADY_LOGGED_IN) from the different-user-
+ * type case (CKR_USER_ANOTHER_ALREADY_LOGGED_IN).
  */
 
 #ifdef HAVE_CONFIG_H
@@ -49,6 +46,7 @@
 #endif
 
 #include "testdata.h"
+#include "pkcs11_test_util.h"
 
 #define TEST_DIR "./store/login_another_user_test"
 #define WOLFPKCS11_TOKEN_FILENAME "wp11_token_0000000000000001"
@@ -57,66 +55,6 @@ static byte* soPin = (byte*)"password123456";
 static CK_ULONG soPinLen = 14;
 static byte* userPin = (byte*)"wolfpkcs11-test";
 static CK_ULONG userPinLen = 15;
-
-static int test_passed = 0;
-static int test_failed = 0;
-
-#define CHECK_RV(rv, op, expected) do {                                       \
-    if ((rv) != (expected)) {                                                 \
-        fprintf(stderr, "FAIL: %s: expected 0x%lx, got 0x%lx\n", op,          \
-                (unsigned long)(expected), (unsigned long)(rv));              \
-        test_failed++;                                                        \
-    } else {                                                                  \
-        printf("PASS: %s\n", op);                                             \
-        test_passed++;                                                        \
-    }                                                                         \
-} while (0)
-
-#ifndef HAVE_PKCS11_STATIC
-static void* dlib;
-#endif
-static CK_FUNCTION_LIST* funcList;
-
-static CK_RV pkcs11_load(void)
-{
-    CK_RV ret;
-#ifndef HAVE_PKCS11_STATIC
-    CK_C_GetFunctionList func;
-
-    dlib = dlopen(WOLFPKCS11_DLL_FILENAME, RTLD_NOW | RTLD_LOCAL);
-    if (dlib == NULL) {
-        fprintf(stderr, "dlopen error: %s\n", dlerror());
-        return CKR_GENERAL_ERROR;
-    }
-    func = (CK_C_GetFunctionList)dlsym(dlib, "C_GetFunctionList");
-    if (func == NULL) {
-        fprintf(stderr, "Failed to get function list function\n");
-        dlclose(dlib);
-        return CKR_GENERAL_ERROR;
-    }
-    ret = func(&funcList);
-    if (ret != CKR_OK) {
-        dlclose(dlib);
-        return ret;
-    }
-#else
-    ret = C_GetFunctionList(&funcList);
-    if (ret != CKR_OK)
-        return ret;
-#endif
-    return CKR_OK;
-}
-
-static void pkcs11_unload(void)
-{
-#ifndef HAVE_PKCS11_STATIC
-    if (dlib != NULL) {
-        dlclose(dlib);
-        dlib = NULL;
-    }
-#endif
-    funcList = NULL;
-}
 
 static void cleanup_store(void)
 {
@@ -141,6 +79,7 @@ static int run_test(void)
     XMEMSET(label, ' ', sizeof(label));
     XMEMCPY(label, "another-user-test", 17);
 
+    /* Start from an uninitialized token. */
     cleanup_store();
 
     rv = pkcs11_load();
@@ -161,7 +100,7 @@ static int run_test(void)
         goto out;
     slot = slotList[0];
 
-    /* Initialize the token (sets the SO PIN) before opening a session. */
+    /* C_InitToken (sets the SO PIN) requires no open session. */
     rv = funcList->C_InitToken(slot, soPin, soPinLen, label);
     CHECK_RV(rv, "C_InitToken", CKR_OK);
     if (rv != CKR_OK)
@@ -172,7 +111,7 @@ static int run_test(void)
     if (rv != CKR_OK)
         goto out;
 
-    /* --- SO logged in: a second SO login is the same type --- */
+    /* SO logged in: a second SO login is the same user type. */
     rv = funcList->C_Login(session, CKU_SO, soPin, soPinLen);
     CHECK_RV(rv, "C_Login(SO)", CKR_OK);
 
@@ -180,19 +119,18 @@ static int run_test(void)
     CHECK_RV(rv, "C_Login(SO) again -> ALREADY_LOGGED_IN",
              CKR_USER_ALREADY_LOGGED_IN);
 
-    /* USER login while SO is logged in is a *different* type. */
     rv = funcList->C_Login(session, CKU_USER, userPin, userPinLen);
     CHECK_RV(rv, "C_Login(USER) while SO -> ANOTHER_ALREADY_LOGGED_IN",
              CKR_USER_ANOTHER_ALREADY_LOGGED_IN);
 
-    /* Set the user PIN (requires SO login) so phase B can log in as user. */
+    /* Set the user PIN (requires SO login) for the second phase. */
     rv = funcList->C_InitPIN(session, userPin, userPinLen);
     CHECK_RV(rv, "C_InitPIN", CKR_OK);
 
     rv = funcList->C_Logout(session);
     CHECK_RV(rv, "C_Logout(SO)", CKR_OK);
 
-    /* --- USER logged in: a second USER login is the same type --- */
+    /* USER logged in: a second USER login is the same user type. */
     rv = funcList->C_Login(session, CKU_USER, userPin, userPinLen);
     CHECK_RV(rv, "C_Login(USER)", CKR_OK);
 
@@ -200,7 +138,6 @@ static int run_test(void)
     CHECK_RV(rv, "C_Login(USER) again -> ALREADY_LOGGED_IN",
              CKR_USER_ALREADY_LOGGED_IN);
 
-    /* SO login while USER is logged in is a *different* type. */
     rv = funcList->C_Login(session, CKU_SO, soPin, soPinLen);
     CHECK_RV(rv, "C_Login(SO) while USER -> ANOTHER_ALREADY_LOGGED_IN",
              CKR_USER_ANOTHER_ALREADY_LOGGED_IN);
@@ -226,14 +163,5 @@ int main(int argc, char* argv[])
 
     printf("=== wolfPKCS11 C_Login another-user test ===\n");
     run_test();
-
-    printf("\n=== Test Results ===\n");
-    printf("Tests passed: %d\n", test_passed);
-    printf("Tests failed: %d\n", test_failed);
-    if (test_failed == 0)
-        printf("ALL TESTS PASSED!\n");
-    else
-        printf("SOME TESTS FAILED!\n");
-
-    return (test_failed == 0) ? 0 : 1;
+    return pkcs11_test_summary();
 }
