+++ tests/trust_attr_bufsize_test.c
@@ -18,11 +18,9 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
  *
- * Regression test for issue F-4060. GetTrustAttr overwrote the caller's *len
- * with sizeof(CK_ULONG)/sizeof(CK_BBOOL) before delegating to GetULong/GetBool,
- * defeating their buffer-too-small check and overflowing an undersized caller
- * buffer for the NSS trust ULONG/BOOL attributes. C_GetAttributeValue must
- * report CKR_BUFFER_TOO_SMALL instead. NSS-only; skipped otherwise.
+ * Regression test for issue F-4060: reading an NSS trust ULONG/BOOL attribute
+ * (CKA_TRUST_*) into an undersized buffer must report CKR_BUFFER_TOO_SMALL
+ * rather than overflowing it. NSS-only; skipped otherwise.
  */
 
 #ifdef HAVE_CONFIG_H
@@ -48,88 +46,14 @@
 #endif
 
 #include "testdata.h"
+#include "pkcs11_test_util.h"
 
 #define TEST_DIR "./store/trust_attr_bufsize_test"
-
-static int test_passed = 0;
-static int test_failed = 0;
-
-#define CHECK_RV(rv, op, expected) do {                                       \
-    if ((rv) != (expected)) {                                                 \
-        fprintf(stderr, "FAIL: %s: expected 0x%lx, got 0x%lx\n", op,          \
-                (unsigned long)(expected), (unsigned long)(rv));              \
-        test_failed++;                                                        \
-    } else {                                                                  \
-        printf("PASS: %s\n", op);                                             \
-        test_passed++;                                                        \
-    }                                                                         \
-} while (0)
-
-#define CHECK_TRUE(cond, op) do {                                             \
-    if (!(cond)) {                                                            \
-        fprintf(stderr, "FAIL: %s\n", op);                                    \
-        test_failed++;                                                        \
-    } else {                                                                  \
-        printf("PASS: %s\n", op);                                             \
-        test_passed++;                                                        \
-    }                                                                         \
-} while (0)
-
-#ifndef HAVE_PKCS11_STATIC
-static void* dlib;
-#endif
-static CK_FUNCTION_LIST* funcList;
-
-static CK_RV pkcs11_load(void)
-{
-    CK_RV ret;
-#ifndef HAVE_PKCS11_STATIC
-    CK_C_GetFunctionList func;
-
-    dlib = dlopen(WOLFPKCS11_DLL_FILENAME, RTLD_NOW | RTLD_LOCAL);
-    if (dlib == NULL) {
-        fprintf(stderr, "dlopen error: %s\n", dlerror());
-        return CKR_GENERAL_ERROR;
-    }
-    func = (CK_C_GetFunctionList)dlsym(dlib, "C_GetFunctionList");
-    if (func == NULL) {
-        fprintf(stderr, "Failed to get function list function\n");
-        dlclose(dlib);
-        return CKR_GENERAL_ERROR;
-    }
-    ret = func(&funcList);
-    if (ret != CKR_OK) {
-        dlclose(dlib);
-        return ret;
-    }
-#else
-    ret = C_GetFunctionList(&funcList);
-    if (ret != CKR_OK)
-        return ret;
-#endif
-    return CKR_OK;
-}
-
-static void pkcs11_unload(void)
-{
-#ifndef HAVE_PKCS11_STATIC
-    if (dlib != NULL) {
-        dlclose(dlib);
-        dlib = NULL;
-    }
-#endif
-    funcList = NULL;
-}
 
 static int run_test(void)
 {
     CK_RV rv;
-    CK_C_INITIALIZE_ARGS args;
-    CK_SLOT_ID slotList[16];
-    CK_ULONG slotCount = sizeof(slotList) / sizeof(slotList[0]);
-    CK_SLOT_ID slot = 0;
     CK_SESSION_HANDLE session = 0;
-    int sessFlags = CKF_SERIAL_SESSION | CKF_RW_SESSION;
 #ifdef WOLFPKCS11_NSS
     CK_OBJECT_HANDLE obj = CK_INVALID_HANDLE;
     CK_OBJECT_CLASS trustClass = CKO_NSS_TRUST;
@@ -164,21 +88,8 @@ static int run_test(void)
     if (rv != CKR_OK)
         return -1;
 
-    XMEMSET(&args, 0, sizeof(args));
-    args.flags = CKF_OS_LOCKING_OK;
-    rv = funcList->C_Initialize(&args);
-    CHECK_RV(rv, "C_Initialize", CKR_OK);
-    if (rv != CKR_OK)
-        goto out;
-
-    rv = funcList->C_GetSlotList(CK_TRUE, slotList, &slotCount);
-    CHECK_RV(rv, "C_GetSlotList", CKR_OK);
-    if (rv != CKR_OK || slotCount == 0)
-        goto out;
-    slot = slotList[0];
-
-    rv = funcList->C_OpenSession(slot, sessFlags, NULL, NULL, &session);
-    CHECK_RV(rv, "C_OpenSession", CKR_OK);
+    rv = pkcs11_open_session(&session);
+    CHECK_RV(rv, "open session", CKR_OK);
     if (rv != CKR_OK)
         goto out;
 
@@ -201,9 +112,8 @@ static int run_test(void)
     CHECK_TRUE(getAttr.ulValueLen == sizeof(CK_ULONG),
                "size query reports sizeof(CK_ULONG)");
 
-    /* Undersized buffer: claim only 4 bytes and guard the rest. Must report
-     * CKR_BUFFER_TOO_SMALL without writing past the 4 bytes. Pre-fix this
-     * wrote sizeof(CK_ULONG) bytes, overflowing the buffer. */
+    /* Undersized buffer: claim 4 bytes, guard the rest with a canary. Must
+     * report CKR_BUFFER_TOO_SMALL without writing past the 4 bytes. */
     XMEMSET(buffer, 0xAB, sizeof(buffer));
     getAttr.type = CKA_TRUST_SERVER_AUTH;
     getAttr.pValue = buffer;
@@ -217,7 +127,7 @@ static int run_test(void)
     CHECK_TRUE(i == (int)sizeof(buffer),
                "no write past the caller-declared buffer length");
 
-    /* The STEP_UP boolean attribute has the same flaw via GetBool. */
+    /* Same check for the STEP_UP boolean attribute (GetBool path). */
     XMEMSET(buffer, 0xAB, sizeof(buffer));
     getAttr.type = CKA_TRUST_STEP_UP_APPROVED;
     getAttr.pValue = buffer;
@@ -256,14 +166,5 @@ int main(int argc, char* argv[])
 
     printf("=== wolfPKCS11 NSS trust attribute buffer-size test ===\n");
     run_test();
-
-    printf("\n=== Test Results ===\n");
-    printf("Tests passed: %d\n", test_passed);
-    printf("Tests failed: %d\n", test_failed);
-    if (test_failed == 0)
-        printf("ALL TESTS PASSED!\n");
-    else
-        printf("SOME TESTS FAILED!\n");
-
-    return (test_failed == 0) ? 0 : 1;
+    return pkcs11_test_summary();
 }
