+++ tests/verify_recover_class_test.c
@@ -18,10 +18,8 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
  *
- * Regression test for issue F-4067. When C_VerifyRecoverInit resolves a
- * handle to a valid object whose CKA_CLASS is not CKO_PUBLIC_KEY it used to
- * return CKR_KEY_HANDLE_INVALID (reserved for handles that do not refer to
- * keys). A key of the wrong object class must report
+ * Regression test for issue F-4067: a handle that resolves to a valid object
+ * of the wrong CKA_CLASS for C_VerifyRecoverInit must report
  * CKR_KEY_TYPE_INCONSISTENT.
  */
 
@@ -48,78 +46,14 @@
 #endif
 
 #include "testdata.h"
+#include "pkcs11_test_util.h"
 
 #define TEST_DIR "./store/verify_recover_class_test"
-
-static int test_passed = 0;
-static int test_failed = 0;
-
-#define CHECK_RV(rv, op, expected) do {                                       \
-    if ((rv) != (expected)) {                                                 \
-        fprintf(stderr, "FAIL: %s: expected 0x%lx, got 0x%lx\n", op,          \
-                (unsigned long)(expected), (unsigned long)(rv));              \
-        test_failed++;                                                        \
-    } else {                                                                  \
-        printf("PASS: %s\n", op);                                             \
-        test_passed++;                                                        \
-    }                                                                         \
-} while (0)
-
-#ifndef HAVE_PKCS11_STATIC
-static void* dlib;
-#endif
-static CK_FUNCTION_LIST* funcList;
-
-static CK_RV pkcs11_load(void)
-{
-    CK_RV ret;
-#ifndef HAVE_PKCS11_STATIC
-    CK_C_GetFunctionList func;
-
-    dlib = dlopen(WOLFPKCS11_DLL_FILENAME, RTLD_NOW | RTLD_LOCAL);
-    if (dlib == NULL) {
-        fprintf(stderr, "dlopen error: %s\n", dlerror());
-        return CKR_GENERAL_ERROR;
-    }
-    func = (CK_C_GetFunctionList)dlsym(dlib, "C_GetFunctionList");
-    if (func == NULL) {
-        fprintf(stderr, "Failed to get function list function\n");
-        dlclose(dlib);
-        return CKR_GENERAL_ERROR;
-    }
-    ret = func(&funcList);
-    if (ret != CKR_OK) {
-        dlclose(dlib);
-        return ret;
-    }
-#else
-    ret = C_GetFunctionList(&funcList);
-    if (ret != CKR_OK)
-        return ret;
-#endif
-    return CKR_OK;
-}
-
-static void pkcs11_unload(void)
-{
-#ifndef HAVE_PKCS11_STATIC
-    if (dlib != NULL) {
-        dlclose(dlib);
-        dlib = NULL;
-    }
-#endif
-    funcList = NULL;
-}
 
 static int run_test(void)
 {
     CK_RV rv;
-    CK_C_INITIALIZE_ARGS args;
-    CK_SLOT_ID slotList[16];
-    CK_ULONG slotCount = sizeof(slotList) / sizeof(slotList[0]);
-    CK_SLOT_ID slot = 0;
     CK_SESSION_HANDLE session = 0;
-    int sessFlags = CKF_SERIAL_SESSION | CKF_RW_SESSION;
     CK_MECHANISM mech;
     CK_OBJECT_HANDLE secret = CK_INVALID_HANDLE;
     CK_OBJECT_CLASS secretClass = CKO_SECRET_KEY;
@@ -132,7 +66,7 @@ static int run_test(void)
         { CKA_KEY_TYPE,    &genericType,  sizeof(genericType)  },
         { CKA_VALUE,       keyData,       sizeof(keyData)      },
         { CKA_EXTRACTABLE, &ckTrue,       sizeof(ckTrue)       },
-        /* Public object so it is reachable without C_Login. */
+        /* Public so the object is reachable without C_Login. */
         { CKA_PRIVATE,     &ckFalse,      sizeof(ckFalse)      },
     };
     CK_ULONG secretTmplCnt = sizeof(secretTmpl) / sizeof(*secretTmpl);
@@ -142,21 +76,8 @@ static int run_test(void)
     if (rv != CKR_OK)
         return -1;
 
-    XMEMSET(&args, 0, sizeof(args));
-    args.flags = CKF_OS_LOCKING_OK;
-    rv = funcList->C_Initialize(&args);
-    CHECK_RV(rv, "C_Initialize", CKR_OK);
-    if (rv != CKR_OK)
-        goto out;
-
-    rv = funcList->C_GetSlotList(CK_TRUE, slotList, &slotCount);
-    CHECK_RV(rv, "C_GetSlotList", CKR_OK);
-    if (rv != CKR_OK || slotCount == 0)
-        goto out;
-    slot = slotList[0];
-
-    rv = funcList->C_OpenSession(slot, sessFlags, NULL, NULL, &session);
-    CHECK_RV(rv, "C_OpenSession", CKR_OK);
+    rv = pkcs11_open_session(&session);
+    CHECK_RV(rv, "open session", CKR_OK);
     if (rv != CKR_OK)
         goto out;
 
@@ -168,9 +89,7 @@ static int run_test(void)
     XMEMSET(&mech, 0, sizeof(mech));
     mech.mechanism = CKM_RSA_PKCS;
 
-    /* The handle resolves to a secret key, not a public key. Wrong object
-     * class for verify-recover must report CKR_KEY_TYPE_INCONSISTENT, not
-     * CKR_KEY_HANDLE_INVALID. */
+    /* A secret key is the wrong object class for verify-recover. */
     rv = funcList->C_VerifyRecoverInit(session, &mech, secret);
     CHECK_RV(rv, "C_VerifyRecoverInit(secret key handle)",
              CKR_KEY_TYPE_INCONSISTENT);
@@ -194,14 +113,5 @@ int main(int argc, char* argv[])
 
     printf("=== wolfPKCS11 C_VerifyRecoverInit wrong-class test ===\n");
     run_test();
-
-    printf("\n=== Test Results ===\n");
-    printf("Tests passed: %d\n", test_passed);
-    printf("Tests failed: %d\n", test_failed);
-    if (test_failed == 0)
-        printf("ALL TESTS PASSED!\n");
-    else
-        printf("SOME TESTS FAILED!\n");
-
-    return (test_failed == 0) ? 0 : 1;
+    return pkcs11_test_summary();
 }
